pyproject.toml
[tool.poetry]
name = "lyra-trading"
version = "0.1.0"
description = ""
authors = ["8baller <8ball030@gmail.com>"]
readme = "README.md"
packages = [{include = "lyra"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<=3.11.5"
requests = "^2.31.0"
web3 = "^5"
websocket-client = "^1.6.4"
setuptools = "^68.2.2"
rich-click = "^1.7.1"
python-dotenv = ">=0.14.0,<0.18.0"


[tool.poetry.scripts]
lyra= "lyra.cli:cli"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true
# you can skip files as below
#skip_glob = docs/conf.py

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py37', 'py38', 'py39']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
